Implemented EnemyMovement, BehaviorCondition, BehaviorClient
Added the EnemyMovement component, work like the same we made in class, but using the NavMesh.

The behaviorTreeClient is a first prototype of a behavior for the client.

Now the behaviorPolice can shoot (Bullet had to be implemented), and have a different distance for follow end shoot.

Now the condition can be implemented by class (like ICode).

For the next week i would implement a character for test different situation.