Adding V1.0 of Combat System
This Combat System include:

IDamager, an interface to identify who deal the damage;
IDamageble, an interface that identify a damage type and implement a "TakeDamage" method;
DamageContainer, a container of SerializeField useful for the "Attack", like the damager, the type, the damage etc. etc.
HealthModule, a script that keep in mind a CurrentHP and a maxHP of an object, with 3 public methods
GenericDamageContainer, a simple class that use IDamager and on trigger call the TakeDamage if an object is Damageble